+++ packages/protocol-http/src/index.ts
@@ -1,3 +1,4 @@
 export * from "./httpResponse";
 export * from "./httpRequest";
 export * from "./httpHandler";
+export * from "./validHostname";
