--- packages/util-valid-hostname/.npmignore
@@ -1,13 +1,0 @@
-/src/
-/coverage/
-/docs/
-tsconfig.test.json
-*.tsbuildinfo
-
-*.spec.js
-*.spec.d.ts
-*.spec.js.map
-
-*.fixture.js
-*.fixture.d.ts
-*.fixture.js.map
