--- packages/util-valid-hostname/README.md
@@ -1,4 +1,0 @@
-# util-valid-hostname
-
-[![NPM version](https://img.shields.io/npm/v/@aws-sdk/util-valid-hostname/preview.svg)](https://www.npmjs.com/package/@aws-sdk/util-valid-hostname)
-[![NPM downloads](https://img.shields.io/npm/dm/@aws-sdk/util-valid-hostname.svg)](https://www.npmjs.com/package/@aws-sdk/util-valid-hostname)
